Increases size of compress bound return value
Updates the `zxc_compress_bound` function to return a `uint64_t` instead of a `size_t` to accommodate larger input sizes and prevent potential overflow issues on platforms with smaller `size_t` implementations. This change also affects the `zxc_get_decompressed_size` function, updating its return type to `uint64_t` as well.

+++ src/cli/main.c
@@ -739,7 +739,7 @@ int main(int argc, char** argv) {
         double dt_c = zxc_now() - t0;
         fclose(fm);
 
-        size_t max_c = zxc_compress_bound(in_size);
+        uint64_t max_c = zxc_compress_bound(in_size);
         c_dat = malloc(max_c);
         if (!c_dat) goto bench_cleanup;
